Allow the program to enter a REPL
Enter a REPL when the program is run without arguments.

I noticed an issue where the prompt would be printed at the same
time as the tokens. Using the 'println!' macro would print the
prompt before the program waited for user input, but would also
print a new line. The prompt is removed until a fix can be found.