Design tests for easy extension as supported expressions are added
Test design now allows for easy extension as the supported Ruby
expressions grow. New expressions can be appended to the input and
the expected results can be appended to the test conditions.

The test is designed to fail on a failed condition, allowing easy
identification of the failed test case.