Allow tokenization of illegal characters
Previously when the lexer attempted to tokenize an illegal
character the program would panic and exit. An illegal character
is any character not recognized as a part of the language while
tokenizing.

Tokenizing is the act of producing tokens, therefore spotting
syntax errors is not the responsibility of the lexer. Here the
changes instead allow the lexer to produce an illegal token type
with the literal value of the illegal character.