Fix bug skipping characters immediately after numbers
Previously a character immediately after a number would be skipped
by the lexer, resulting in no token being produced for the skipped
character. This was caused by a redundant call to next()
immediately after a number was read. This bug would also not be
detected in the existing test if the last character was preceded by
a number.

Here the changes fix the bug, and test that the number of tests run
matches the number of expected conditions.