Redefine Lexer as an Iterator
Previously Lexer was a Trait which could be implemented on any
type suitable for lexical analysis e.g. String.

Here the changes redefine Lexer as an Iterator. Implementing the
Iterator trait allows Lexer to produce a Token on demand, where
previously the entire input was tokenized up front and output
as a Vector of Tokens. This allows for better integration with
the chosen design of the Parser.